Generate JSON output from corresponding templates in the theme
Generate a header, the pre app content, the post app content, and
a footer to pass as pre and post content in a JSON object. Themes
can use web-template-pre.php and web-template-post.php for basic
templating. Specific slugs can be used as well to target apps that
may have different design requirements.